requirements.txt
anthropic
chromadb
streamlit
requests
pyyaml
python-dotenv